requirements.txt
django
django-bakery